Resolved conflicts by accepting main's version for:
- arithmetic.rs: Keep Float for integer division (separate fix in PR #915)
- nonagg.rs: Use main's implementation

+++ crates/executor/src/evaluator/operators/arithmetic.rs
@@ -215,15 +215,15 @@ impl ArithmeticOps {
         use SqlValue::*;
 
         match (left, right) {
-            // Integer division with zero check - returns Numeric to preserve precision
+            // Integer division with zero check - returns Float to preserve precision
             (Integer(a), Integer(b)) => {
                 if *b == 0 {
                     return Err(ExecutorError::DivisionByZero);
                 }
-                Ok(Numeric(*a as f64 / *b as f64))
+                Ok(Float((*a as f64 / *b as f64) as f32))
             }
 
-            // Mixed exact numeric types - promote to i64
+            // Mixed exact numeric types - promote to i64, return Float for precision
             (left_val, right_val)
                 if is_exact_numeric(left_val) && is_exact_numeric(right_val) =>
             {
@@ -232,7 +232,7 @@ impl ArithmeticOps {
                 if right_i64 == 0 {
                     return Err(ExecutorError::DivisionByZero);
                 }
-                Ok(Integer(left_i64 / right_i64))
+                Ok(Float((left_i64 as f64 / right_i64 as f64) as f32))
             }
 
             // Approximate numeric types - promote to f64
